Rename modules folder to 'core' and add 'plugins'
